fix: rename copied readmes to about.md
